Still need improvement!
Have the CLI version corrected where the application for CLI version gives the correct results.  In this CLI version, if a "A" player sets up the ships locations via the text file, the system go and read that text file and then memorized the ships locations and it assigns inside the game. Then, when "B" player plays the board game, the application will only allow "Hit" and "Miss" to the coordinates where player "A" assigned earlier. However, this function currently is only working in the CLI side but not on the GUI side.